Message validation is being sent over now. Wasn't happening before
Needs to be rethought a bit because in some places the errors are a list
and others they are a single item...
